+++ app/Http/Controllers/SurveyController.php
@@ -182,96 +182,6 @@ class SurveyController extends Controller
         return redirect()->route('listsurvey', [$clientSlug, $projectSlug])->with('success', 'Update successfully.');
     }
 
-    public function submission(Survey $surveyModel, $clientSlug, $projectSlug, $id)
-    {
-        // Fetch survey, questions, project, client
-        $survey = Survey::findOrFail($id);
-        $project = DB::table('projects')->where('slug', $projectSlug)->get();
-        $client = DB::table('clients')->where('slug', $clientSlug)->get();
-        $page = QuestionPage::where('survey_id', $id)->get();
-        // Prepare data to pass to the view
-        $res = Response::where('survey_id', $id)->where('user_id', Auth::user()->id)->first();
-        $flow = Flow::where('survey_id', $id)->get();
-        $formattedPage = $page->map(function ($p) {
-            return [
-                'id' => $p->id,
-                'page_name' => $p->page_name,
-                'survey_id' => $p->survey_id,
-                'flow' => $p->flow->map(function ($f) {
-                    return ['flow'=>$f];
-                }),
-                'question' => $p->question->map(function ($q) {
-                    return [
-                        'id' => $q->id,
-                        'question_text' => $q->question_text,
-                        'question_type_id' => $q->question_type_id,
-                        'survey_id' => $q->survey_id,
-                        'order' => $q->order,
-                        'required' => $q->required,
-                        'choice' => $q->choice,
-                    ];
-                }),
-            ];
-        });
-
-        // Render the view
-        return Inertia::render(
-            'Client/Projects/Surveys/SubmissionSurvey',
-            [
-                'surveys' => $survey,
-                'projects' => $project,
-                'clients' => $client,
-                'page' => $formattedPage,
-                'pagee' => $page,
-                'flow' => $flow
-                // 'responses' => $res['id'] ?? null
-            ]
-        );
-    }
-
-    public function report(Survey $surveyModel, $clientSlug, $projectSlug, $surveyId, $responseId)
-    {
-        $survey = Survey::findOrFail($surveyId);
-        $response = Response::with('user')->where('survey_id', $surveyId)->findOrFail($responseId);
-        $project = DB::table('projects')->where('slug', $projectSlug)->get();
-        $client = DB::table('clients')->where('slug', $clientSlug)->get();
-        $answers = Answer::where('response_id', $responseId)->get();
-        $bio = $response->user->biodata;
-        $page = QuestionPage::where('survey_id', $surveyId)->get();
-        // Prepare data to pass to the view
-        $formattedPage = $page->map(function ($p) {
-            return [
-                'id' => $p->id,
-                'page_name' => $p->page_name,
-                'survey_id' => $p->survey_id,
-                'question' => $p->question->map(function ($q) {
-                    return [
-                        'id' => $q->id,
-                        'question_text' => $q->question_text,
-                        'question_type_id' => $q->question_type_id,
-                        'survey_id' => $q->survey_id,
-                        'order' => $q->order,
-                        'required' => $q->required,
-                        'choice' => $q->choice,
-                        'answer' => $q->answer
-                    ];
-                }),
-            ];
-        });
-        return Inertia::render(
-            'Client/Projects/Surveys/ReportSurvey',
-            [
-                'surveys' => $survey,
-                'projects' => $project,
-                'clients' => $client,
-                'page' => $formattedPage,
-                'responses' => $response,
-                'answer' => $answers,
-                'biodata' => $bio,
-            ]
-        );
-    }
-
     public function destroy($clientSlug, $projectSlug, $id)
     {
         $survey =  Survey::findOrFail($id);
